backend/django/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
pillow==11.1.0
psycopg==3.2.3
pycparser==2.22
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
